calc/requirements.txt
altgraph==0.17
app==0.0.1
asgiref==3.3.1
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
csvmapper==0.7
decorator==4.4.1
Django==3.1.4
django-bootstrap3==14.2.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-sitetree==1.16.0
django-tables2==2.3.3
django-widget-tweaks==1.4.8
dnspython==1.16.0
easygui==0.98.1
fabric==2.5.0
ftputil==3.4
future==0.18.2
idna==2.10
invoke==1.4.1
mysql-connector-python==8.0.19
numpy==1.18.1
pandas==1.0.1
paramiko==2.7.1
pefile==2019.4.18
protobuf==3.6.1
psutil==5.7.2
public==2019.4.13
py2exe==0.9.2.2
pycparser==2.19
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
PyNaCl==1.3.0
pyslack==0.5.0
python-dateutil==2.8.1
pytz==2019.3
pywin32-ctypes==0.2.0
requests==2.24.0
selenium==3.141.0
self==2019.4.13
six==1.14.0
slack-notifications==0.0.7
sqlparse==0.4.1
UNKNOWN==0.0.0
urllib3==1.25.8
wget==3.2
xlrd==1.2.0
psycopg2==2.5.4

